+++ Changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased: 5.1.1-rc.2]
+### Changed
+ - FIO-9992: check for server when triggering the captcha
+
 ## 5.1.0
 ### Changed
  - Official Release